trading-agent/requirements.txt
# Core LangGraph and LangChain
langgraph==0.2.45
langchain==0.3.7
langchain-openai==0.2.8

# Trading APIs
alpaca-py==0.28.3
yfinance==0.2.49

# Data Processing
pandas==2.2.3
numpy==1.26.4
ta==0.11.0  # Technical Analysis library

# Database
sqlalchemy==2.0.36

# Configuration
python-dotenv==1.0.1

# Scheduling (for live trading)
schedule==1.2.2

# Utilities
requests==2.32.3
pytz==2024.2
